Extract container plugin
