+++ plugins/provisioners/podman/cap/linux/podman_installed.rb
@@ -4,20 +4,7 @@ module VagrantPlugins
       module Linux
         module PodmanInstalled
           def self.podman_installed(machine)
-            paths = [
-              "/bin/podman",
-              "/usr/bin/podman",
-              "/usr/local/bin/podman",
-              "/usr/sbin/podman",
-            ]
-
-            paths.each do |p|
-              if machine.communicate.test("test -f #{p}", sudo: true)
-                return true
-              end
-            end
-
-            return false
+            machine.communicate.test("command -v podman")
           end
         end
       end
