+++ plugins/provisioners/podman/cap/redhat/podman_install.rb
@@ -9,7 +9,7 @@ module VagrantPlugins
             case machine.guest.capability("flavor")
             when :rhel_7
               machine.communicate.tap do |comm|
-                comm.sudo("curl -sSL https://download.opensuse.org/repositories/devel:kubic:libcontainers:stable/CentOS_7/devel:kubic:libcontainers:stable.repo -o /etc/yum.repos.d/")
+                comm.sudo("curl curl -L -o /etc/yum.repos.d/devel:kubic:libcontainers:stable.repo https://download.opensuse.org/repositories/devel:/kubic:/libcontainers:/stable/CentOS_7/devel:kubic:libcontainers:stable.repo")
                 comm.sudo("yum -q -y install podman")
               end
             when :rhel_8
@@ -17,7 +17,7 @@ module VagrantPlugins
                 comm.sudo("dnf -y module disable container-tools")
                 comm.sudo("dnf -y install 'dnf-command(copr)'")
                 comm.sudo("dnf -y copr enable rhcontainerbot/container-selinux")
-                comm.sudo("curl -sSL https://download.opensuse.org/repositories/devel:kubic:libcontainers:stable/CentOS_8/devel:kubic:libcontainers:stable.repo -o /etc/yum.repos.d/")
+                comm.sudo("curl -L -o /etc/yum.repos.d/devel:kubic:libcontainers:stable.repo https://download.opensuse.org/repositories/devel:/kubic:/libcontainers:/stable/CentOS_8/devel:kubic:libcontainers:stable.repo")
                 comm.sudo("dnf -y install podman")
               end
             else
